docstore tests
