+++ haystack/preview/__init__.py
@@ -1,2 +1,3 @@
 from canals import node
+from haystack.preview.dataclasses import Document
 from haystack.preview.pipeline import Pipeline, PipelineError, NoSuchStoreError, load_pipelines, save_pipelines
