+++ haystack/preview/dataclasses/__init__.py
@@ -1,0 +1,1 @@
+from haystack.preview.dataclasses.document import Document
