+++ haystack/preview/document_stores/__init__.py
@@ -1,0 +1,2 @@
+from haystack.preview.document_stores.memory.document_store import MemoryDocumentStore
+from haystack.preview.document_stores._utils import StoreError, DuplicateDocumentError, MissingDocumentError
